Cleanup to files: - Fixed resource leaks - Fixed bug in Window class which would narrow down the window resolution values because of wrong data type - Integrated Dear ImGui update routine - Integrated basic rendering routine

+++ src/core/window/window.cpp
@@ -30,12 +30,12 @@ namespace Borealis::Core
 		UpdateWindowInternal();
 	}
 
-	uint8 Window::GetWindowWidth() const
+	uint16 Window::GetWindowWidth() const
 	{
 		return GetWindowWidthInternal();
 	}
 	
-	uint8 Window::GetWindowHeight() const
+	uint16 Window::GetWindowHeight() const
 	{
 		return GetWindowHeightInternal();
 	}
